added RankNet ndcg@5: 0.38609914482977487

+++ dataloader.py
@@ -29,4 +29,4 @@ def load_data(filename):
             y.append(yi)
             q.append(qi)
             info.append(comment)
-        return np.array(x, dtype=np.float32), np.array(y, dtype=np.int), np.array(q, dtype=np.int)
+        return np.array(x, dtype=np.float32), np.array(y, dtype=np.float32), np.array(q, dtype=np.float32)
